move azure/quantum/optimization to azure/quantum/target
