+++ azure-quantum/azure/quantum/optimization/__init__.py
@@ -4,8 +4,8 @@
 # Licensed under the MIT License.
 ##
 
-from .term import *
-from .problem import *
-from .solvers import *
-from .streaming_problem import *
-from .online_problem import *
+from azure.quantum.target.optimization.term import *
+from azure.quantum.target.optimization.problem import *
+from azure.quantum.target.optimization.solvers import *
+from azure.quantum.target.optimization.streaming_problem import *
+from azure.quantum.target.optimization.online_problem import *
