--- azure-quantum/azure/quantum/optimization/oneqbit/__init__.py
@@ -1,7 +1,0 @@
-# coding=utf-8
-##
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-##
-
-from .solvers import *
