+++ azure-quantum/tests/unit/test_streaming_problem.py
@@ -9,17 +9,14 @@
 
 import unittest
 import json
-import pytest
 from typing import List
 
-from azure.quantum import Workspace
-from azure.quantum.optimization import (
+from azure.quantum.target.optimization import (
     StreamingProblem,
     Problem,
     ProblemType,
     Term,
 )
-from azure.quantum.storage import download_blob
 from common import QuantumTestBase
